refactor: deduplicate shared helpers between rebase_pr and claude_step
Move _rebase_onto_target() and _truncate() from rebase_pr.py into
claude_step.py alongside _run_git(), eliminating copy-paste duplication
across pipeline modules. Update test mock paths accordingly.
